pyproject.toml
[project]
name = "crm-backend"
version = "0.1.0"
description = "crm for roofing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "geoalchemy2>=0.18.1",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
